+++ cpp/arduino/Arduino.cpp
@@ -1,8 +1,10 @@
 #include "Arduino.h"
 
-struct unit_test_state godmode = {
-  0, // micros
-};
+GodmodeState godmode = GodmodeState();
+
+GodmodeState* GODMODE() {
+  return &godmode;
+}
 
 unsigned long millis() {
   return godmode.micros / 1000;
@@ -10,4 +12,12 @@ unsigned long millis() {
 
 unsigned long micros() {
   return godmode.micros;
+}
+
+void delay(unsigned long millis) {
+  godmode.micros += millis * 1000;
+}
+
+void delayMicroseconds(unsigned long micros) {
+  godmode.micros += micros;
 }